requirements.txt
fastapi
uvicorn
pymongo
qdrant-client
google-generativeai
python-dotenv
pydantic